requirements.txt
django-filter==2.0.0
djangorestframework==3.8.2
Markdown==2.6.11
django-environ==0.4.5